pyproject.toml
[project]
name = "podcast-examples"
version = "1.0.0"
description = "Podcast examples project with RunningHub API integration"
authors = [
    {name = "User", email = "user@example.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.28.0",
    "python-dotenv>=1.0.0",
    "zai-sdk>=0.0.4.2",
    "lunarcalendar>=0.0.9",
    "pymupdf>=1.24.11",
    "pillow>=10.4.0",
    "poppler-utils>=0.1.0",
    "PyMuPDF>=1.24.11",
    "magika>=1.0.1",
    "markitdown>=0.0.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/podcast_examples"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
